+++ abinois_is_the_true_leader_of_this_group__word/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: abinois <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/02 17:24:14 by abinois           #+#    #+#             */
-/*   Updated: 2019/05/14 11:55:39 by abinois          ###   ########.fr       */
+/*   Updated: 2019/05/15 15:17:15 by abinois          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,8 +94,8 @@ void			ft_putnstr(const char *s, size_t n);
 void			ft_fili(const char *file, size_t line);
 char			*ft_llutoa(unsigned long long nbr);
 char			*ft_lltoa(long long nbr);
-char			*ft_octatoa(unsigned int nbr);
-char			*ft_hexatoa(unsigned int nbr);
+char			*ft_octatoa(unsigned long long nbr);
+char			*ft_hexatoa(unsigned long long nbr);
 char			*ft_str_up(char *s);
 
 #endif
